test(api): move shared scheduler test helpers to common-tournaments
